chore: move files in to subdirectories
