+++ cypress.json
@@ -2,5 +2,5 @@
     "baseUrl": "http://localhost:3000",
     "video": true,
     "projectId": "5fk191",
-    "testFiles": "**/*.feature"
+    "testFiles": "**/*.{feature,features}"
 }
