+++ cypress/integration/common/before_each_visit_root.js
@@ -1,0 +1,5 @@
+import { EXTENDED_TIMEOUT } from '../../support/utils'
+
+beforeEach(() => {
+    cy.visit('/', EXTENDED_TIMEOUT)
+})
