+++ cypress/integration/common/view/dashboard_description.js
@@ -1,5 +1,5 @@
 import { Then } from 'cypress-cucumber-preprocessor/steps'
-import { dashboardDescriptionSel } from '../../selectors/viewDashboard'
+import { dashboardDescriptionSel } from '../../../selectors/viewDashboard'
 
 Then('the dashboard description is not displayed', () => {
     cy.get(dashboardDescriptionSel).should('not.exist')
