+++ cypress/integration/common/view/open_SL_dashboard.js
@@ -1,0 +1,21 @@
+import { Given, Then } from 'cypress-cucumber-preprocessor/steps'
+import { dashboards } from '../../../assets/backends'
+import { EXTENDED_TIMEOUT } from '../../../support/utils'
+import { chartSel } from '../../../selectors/dashboardItem'
+import {
+    dashboardTitleSel,
+    dashboardChipSel,
+} from '../../../selectors/viewDashboard'
+
+Given('I open the {string} dashboard', title => {
+    cy.get(dashboardChipSel, EXTENDED_TIMEOUT).contains(title).click()
+})
+
+Then('the {string} dashboard displays in view mode', title => {
+    cy.location().should(loc => {
+        expect(loc.hash).to.equal(dashboards[title].route)
+    })
+
+    cy.get(dashboardTitleSel).should('be.visible').and('contain', title)
+    cy.get(chartSel, EXTENDED_TIMEOUT).should('exist')
+})
