+++ cypress/integration/common/view/toggle_show_more_dashboards.js
