+++ cypress/integration/edit/edit_dashboard.feature
@@ -5,7 +5,7 @@ Feature: Creating, editing and deleting dashboard
         Given I choose to create new dashboard
         And dashboard title is added
         And dashboard items are added
-        And I click outside menu
+        And I close the item selector
         And dashboard is saved
         Then the dashboard displays in view mode
         And the saved dashboard should be displayed
@@ -31,7 +31,7 @@ Feature: Creating, editing and deleting dashboard
         Given I open existing dashboard
         When I choose to edit dashboard
         And dashboard items are added
-        And I click outside menu
+        And I close the item selector
         And I click Exit without saving
         And I decide to continue editing
         Then the dashboard displays in edit mode
@@ -69,7 +69,6 @@ Feature: Creating, editing and deleting dashboard
     Scenario: I move an item on a dashboard
         Given I open existing dashboard
         When I choose to edit dashboard
-        Then the dashboard displays in edit mode
         And the chart item is displayed
         Then no analytics requests are made when item is moved
 
@@ -104,7 +103,6 @@ Feature: Creating, editing and deleting dashboard
         When I click to star the dashboard
         Then the dashboard is starred
         When I choose to edit dashboard
-        Then the dashboard displays in edit mode
         And dashboard is saved
         Then the dashboard displays in view mode
         And the dashboard is starred
