+++ cypress/integration/edit/edit_dashboard/sharing.js
