+++ cypress/integration/edit/edit_dashboard/show_description.js
