+++ cypress/integration/edit/edit_dashboard/star_dashboard.js
