+++ cypress/integration/edit/edit_dashboard/translate_dashboard.js
