+++ cypress/integration/edit/edit_errors/error_during_delete.js
