+++ cypress/integration/edit/filter_restrict.feature
@@ -54,7 +54,7 @@ Feature: Editing Filter Restrictions
         And I click Exit without saving
         And I confirm I want to discard changes
         Then the dashboard displays in view mode
-        When I open Edit mode
+        When I choose to edit dashboard
         And I click on Filter settings
         Then Filter Restrictions are not restricted
 
@@ -80,6 +80,6 @@ Feature: Editing Filter Restrictions
         And I save the dashboard
         Then Add Filter button is not visible
         And the dashboard displays in view mode
-        When I open Edit mode
+        When I choose to edit dashboard
         And I delete the dashboard
         Then different dashboard displays in view mode
