+++ cypress/integration/view/fullscreen.feature
