+++ cypress/integration/view/fullscreen/fullscreen.js
