+++ cypress/integration/view/item_context_menu/item_context_menu.js
