+++ cypress/integration/view/responsive_dashboard.feature
