+++ cypress/integration/view/responsive_dashboard/phone_landscape.js
