+++ cypress/integration/view/view_dashboard.feature
