+++ cypress/integration/view/view_dashboard/control_bar.js
