+++ cypress/integration/view/view_dashboard/dashboard_items_without_shape.js
