+++ cypress/integration/view/view_dashboard/index.js
