+++ cypress/integration/view/view_dashboard/print.js
