+++ cypress/integration/view/view_errors.feature
@@ -1,4 +1,4 @@
-Feature: Error scenarios
+Feature: Errors while in view mode
 
     @nonmutating
     Scenario: There are no dashboards
@@ -33,26 +33,6 @@ Feature: Error scenarios
     #     When I open the "Delivery" dashboard
     #     Then the "Delivery" dashboard displays in view mode
 
-    @nonmutating
-    Scenario: An error occurs while saving a dashboard that I don't have access to
-        Given I open the "Delivery" dashboard
-        When I choose to edit dashboard
-        And I save dashboard that I don't have access to save
-        Then I remain in edit mode and error message is displayed
-
-    @nonmutating
-    Scenario: A 500 error is thrown when I save the dashboard
-        Given I open the "Delivery" dashboard
-        When I choose to edit dashboard
-        And A 500 error is thrown when I save the dashboard
-        Then I remain in edit mode and error message is displayed
-
-    @nonmutating
-    Scenario: A 500 error is thrown when I delete the dashboard
-        Given I open the "Delivery" dashboard
-        When I choose to edit dashboard
-        And A 500 error is thrown when I delete the dashboard
-        Then I remain in edit mode and error message is displayed
 
     @nonmutating
     Scenario: I enter edit mode of a dashboard I do not have access to edit
@@ -75,6 +55,3 @@ Feature: Error scenarios
 #     When clicking to show description fails
 #     Then a warning message is displayed stating that show description failed
 #     And the dashboard description is not displayed
-
-
-
