+++ cypress/integration/view/view_errors/error_while_show_description.js
