+++ cypress/integration/view/view_errors/error_while_starring_dashboard.js
