+++ cypress/integration/view/view_errors/invalid_dashboard_id.js
