+++ cypress/integration/view/view_errors/no_dashboards.js
