+++ cypress/integration/view/view_errors/no_edit_access.js
